move codes for mixed encoding
